Extract add_calibration to review_response_map_controller
